inline Badge docs
