+++ packages/argo-admin-react/src/components/Badge/examples/Badge.example.tsx
@@ -1,0 +1,11 @@
+import React from 'react';
+import {extend, render, Badge} from '@shopify/argo-admin-react';
+
+function App() {
+  return <Badge message="Example message" status="success" />;
+}
+
+extend(
+  'Playground',
+  render(() => <App />),
+);
