+++ packages/argo-admin-react/src/components/Badge/index.ts
@@ -1,0 +1,1 @@
+export {Badge} from './Badge';
