+++ packages/argo-admin/src/components/Badge/Badge.ts
@@ -1,0 +1,19 @@
+import {createRemoteComponent} from '@remote-ui/core';
+
+type Status = 'success' | 'info' | 'attention' | 'warning' | 'new';
+
+export interface BadgeProps {
+  /**
+   * The content to display inside the badge.
+   */
+  message: string;
+  /**
+   * Set the colour of the badge for the given status. Defaults to no status.
+   */
+  status?: Status;
+}
+
+/**
+ * Badges are used to inform merchants of the status of an object, or the status of an action that’s been taken.
+ */
+export const Badge = createRemoteComponent<'Badge', BadgeProps>('Badge');
