+++ packages/argo-admin/src/components/Badge/content/guidelines.md
@@ -1,0 +1,10 @@
+## Guidelines
+
+| ✅ Do                                                                     | 🛑 Don't                                           |
+| ------------------------------------------------------------------------- | -------------------------------------------------- |
+| Use Badges to display the status of an object                             | Use Badges without context, or as singular objects |
+| Use Badges to represent discrete information                              |                                                    |
+| Have clear and helpful relationship to the content or task they represent |                                                    |
+| Badges can be stacked horizontally                                        |                                                    |
+
+For more guidelines, refer to Polaris' [Badge best practices](https://polaris.shopify.com/components/images-and-icons/badge#section-best-practices).
