+++ packages/argo-admin/src/components/Badge/index.ts
@@ -1,0 +1,2 @@
+export {Badge} from './Badge';
+export type {BadgeProps} from './Badge';
